Playbook deletion + Actions column in Plugins page (#1908)
* playbook deletion button + actions column

* prettier

* black

* delete quark-rules

* gitignore

* changes

* black

* isort

* Playbook Ownership (#1916)

* Bump actions/checkout from 3.1.0 to 4.1.0 (#1907)

Bumps [actions/checkout](https://github.com/actions/checkout) from 3.1.0 to 4.1.0.
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/v3.1.0...v4.1.0)

---
updated-dependencies:
- dependency-name: actions/checkout
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Refactor pivot (#1902)

* Refactor pivot


* Fixes


* More


* Linters


* Update api_app/pivots_manager/models.py


* Fix


* Fixes


* Fixes


* Fixes


* Fixes


* Fixes


* Fixes


* Fixes


* Deepsource


* Fix


* fix test frontend

* Prettier


* Fixes


* Fixes


* Blake


* Fix typo


* Fix


* improved pivot frontend

* frontend

* Fix


* Fix url property


* Fix url property


* improved pivoting page

* Removed import


* prettier

* Fixes


* fixed frontend tests

* prettier

* Fix yaraify url


* Adding removal of existent directory


---------


* Bump checkdmarc from 4.7.0 to 4.8.4 in /requirements (#1903)

Bumps [checkdmarc](https://github.com/domainaware/checkdmarc) from 4.7.0 to 4.8.4.
- [Changelog](https://github.com/domainaware/checkdmarc/blob/master/CHANGELOG.md)
- [Commits](https://github.com/domainaware/checkdmarc/commits/4.8.4)

---
updated-dependencies:
- dependency-name: checkdmarc
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump pyelftools from 0.29 to 0.30 in /requirements (#1887)

Bumps [pyelftools](https://github.com/eliben/pyelftools) from 0.29 to 0.30.
- [Changelog](https://github.com/eliben/pyelftools/blob/master/CHANGES)
- [Commits](https://github.com/eliben/pyelftools/compare/v0.29...v0.30)

---
updated-dependencies:
- dependency-name: pyelftools
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump django-storages from 1.13.1 to 1.14 in /requirements (#1877)

Bumps [django-storages](https://github.com/jschneier/django-storages) from 1.13.1 to 1.14.
- [Changelog](https://github.com/jschneier/django-storages/blob/master/CHANGELOG.rst)
- [Commits](https://github.com/jschneier/django-storages/commits)

---
updated-dependencies:
- dependency-name: django-storages
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Bump quark-engine from 23.7.1 to 23.8.1 in /requirements (#1876)

Bumps [quark-engine](https://github.com/quark-engine/quark-engine) from 23.7.1 to 23.8.1.
- [Release notes](https://github.com/quark-engine/quark-engine/releases)
- [Commits](https://github.com/quark-engine/quark-engine/compare/v23.7.1...v23.8.1)

---
updated-dependencies:
- dependency-name: quark-engine
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Split backend from frontend


* Pivots page is shown only when there are pivots


* Generalization of pivot config (#1909)

* Generalization of config


* Fixes


* Typo


---------


* Fix pivot


* Fix serializer


* Update even personal/org plugin config


* Added condition to find previous job


* Ownership to playbooks


* PivotMap can be created without a pivotConfig


* VisaulizerReport has the config field


* Job response can now be "accepted" or "exists": if a new job is actually created or not


* Signals + black + migrations


* More tests + constraints


* Bump dnstwist[full] from 20230509 to 20230918 in /requirements (#1912)

Bumps [dnstwist[full]](https://github.com/elceef/dnstwist) from 20230509 to 20230918.
- [Release notes](https://github.com/elceef/dnstwist/releases)
- [Commits](https://github.com/elceef/dnstwist/commits)

---
updated-dependencies:
- dependency-name: dnstwist[full]
  dependency-type: direct:production
  update-type: version-update:semver-major
...


* Bump django-filter from 23.2 to 23.3 in /requirements (#1913)

Bumps [django-filter](https://github.com/carltongibson/django-filter) from 23.2 to 23.3.
- [Release notes](https://github.com/carltongibson/django-filter/releases)
- [Changelog](https://github.com/carltongibson/django-filter/blob/main/CHANGES.rst)
- [Commits](https://github.com/carltongibson/django-filter/compare/23.2...23.3)

---
updated-dependencies:
- dependency-name: django-filter
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* Fixes


* Start_time is updated on worker execution


* Fix


* Fix


* Fix test


* Fixes


* Fixes


* FIx post merge


---------


* changes

* black

* prettier

* flake

* frontend tests

* prettier

* changes

* fixed migrations + adjusted SaveAsPlaybookButton

* adjested backend tests

* adjusted PlaybooksDeletionButton

---------

+++ frontend/src/components/plugins/utils/utils.jsx
@@ -12,13 +12,13 @@ import {
   UncontrolledPopover,
   Button,
   Collapse,
+  Modal,
+  ModalHeader,
+  ModalBody,
 } from "reactstrap";
 import { RiHeartPulseLine } from "react-icons/ri";
-import {
-  MdInfo,
-  MdOutlineCheckCircleOutline,
-  MdOutlineHideSource,
-} from "react-icons/md";
+import { MdInfo, MdDelete } from "react-icons/md";
+import { BsPeopleFill } from "react-icons/bs";
 
 import { IconButton, BooleanIcon, ArrowToggleIcon } from "@certego/certego-ui";
 
@@ -317,6 +317,7 @@ export function OrganizationPluginStateToggle({
   const {
     isUserOwner,
     organization,
+    noOrg,
     fetchAll: fetchAllOrganizations,
   } = useOrganizationStore(
     React.useCallback(
@@ -324,14 +325,17 @@ export function OrganizationPluginStateToggle({
         fetchAll: state.fetchAll,
         isUserOwner: state.isUserOwner,
         organization: state.organization,
+        noOrg: state.noOrg,
       }),
       [],
     ),
   );
-  let title = disabled ? "Enable" : "Disable";
-  if (!organization.name) title = "You're not a part of any organization";
-  else if (!isUserOwner)
+  let title = disabled
+    ? `Enable ${pluginName} for organization`
+    : `Disable ${pluginName} for organization`;
+  if (!isUserOwner) {
     title = `You're not an owner of your organization - ${organization.name}`;
+  }
   let baseUrl = "";
   if (type === pluginType.ANALYZER) {
     baseUrl = ANALYZERS_CONFIG_URI;
@@ -342,22 +346,26 @@ export function OrganizationPluginStateToggle({
   }
 
   const onClick = async () => {
-    if (disabled) await axios.delete(`${baseUrl}/${pluginName}/organization`);
-    else await axios.post(`${baseUrl}/${pluginName}/organization`);
-    fetchAllOrganizations();
-    refetch();
+    if (isUserOwner) {
+      if (disabled) await axios.delete(`${baseUrl}/${pluginName}/organization`);
+      else await axios.post(`${baseUrl}/${pluginName}/organization`);
+      fetchAllOrganizations();
+      refetch();
+    }
   };
   return (
-    <div className="d-flex flex-column align-items-center">
-      <IconButton
-        id={`table-pluginstatebtn__${pluginName}`}
-        color={disabled ? "danger" : "info"}
-        size="sm"
-        Icon={disabled ? MdOutlineHideSource : MdOutlineCheckCircleOutline}
-        title={title}
-        onClick={onClick}
-        titlePlacement="top"
-      />
+    <div className={`d-flex align-items-center ${noOrg ? "" : "px-2"}`}>
+      {!noOrg && (
+        <IconButton
+          id={`table-pluginstatebtn__${pluginName}`}
+          color={disabled ? "dark" : "success"}
+          size="sm"
+          Icon={BsPeopleFill}
+          title={title}
+          onClick={onClick}
+          titlePlacement="top"
+        />
+      )}
     </div>
   );
 }
@@ -384,6 +392,78 @@ export function PlaybooksCollapse({ value, pluginType_ }) {
           ))}
         </ul>
       </Collapse>
+    </div>
+  );
+}
+
+export function PlaybooksDeletionButton({ playbookName }) {
+  const [showModal, setShowModal] = React.useState(false);
+
+  const { deletePlaybook, retrievePlaybooksConfiguration } =
+    usePluginConfigurationStore(
+      React.useCallback(
+        (state) => ({
+          deletePlaybook: state.deletePlaybook,
+          retrievePlaybooksConfiguration: state.retrievePlaybooksConfiguration,
+        }),
+        [],
+      ),
+    );
+
+  const onClick = async () => {
+    try {
+      await deletePlaybook(playbookName);
+      setShowModal(false);
+      await retrievePlaybooksConfiguration();
+    } catch {
+      // handle error in deletePlaybook
+    }
+  };
+
+  return (
+    <div>
+      <IconButton
+        id={`playbook-deletion-${playbookName}`}
+        color="danger"
+        size="sm"
+        Icon={MdDelete}
+        title="Delete playbook"
+        onClick={() => setShowModal(true)}
+        titlePlacement="top"
+      />
+      <Modal
+        id={`modal-playbook-deletion-${playbookName}`}
+        autoFocus
+        centered
+        zIndex="1050"
+        size="lg"
+        keyboard={false}
+        backdrop="static"
+        labelledBy="Playbook deletion modal"
+        isOpen={showModal}
+      >
+        <ModalHeader className="mx-2" toggle={() => setShowModal(false)}>
+          <small className="text-info">Delete playbook</small>
+        </ModalHeader>
+        <ModalBody className="d-flex justify-content-between my-2 mx-2">
+          <div>
+            Do you want to delete the playbook:{" "}
+            <span className="text-info">{playbookName}</span>?
+          </div>
+          <div className="d-flex justify-content-between">
+            <Button className="mx-2" color="danger" size="sm" onClick={onClick}>
+              Delete
+            </Button>
+            <Button
+              className="mx-2"
+              size="sm"
+              onClick={() => setShowModal(false)}
+            >
+              Cancel
+            </Button>
+          </div>
+        </ModalBody>
+      </Modal>
     </div>
   );
 }
@@ -416,4 +496,8 @@ PluginHealthCheckButton.propTypes = {
 PlaybooksCollapse.propTypes = {
   value: PropTypes.array.isRequired,
   pluginType_: PropTypes.oneOf(["analyzers", "connectors"]).isRequired,
+};
+
+PlaybooksDeletionButton.propTypes = {
+  playbookName: PropTypes.string.isRequired,
 };
